server.js
Description: Assuming an IOT that feeds data readings to redis cache, the following
Node.js code will read from the redis cache and return data to an HTTP client such 
as jQuery. The Redis cache contains a list of units or locations to which devices are
deployed. The devices send data back as a key composed of the device id and the unit id.
The value to the key is a JSON string of measurements, e.g. temperature, lighting, etc.
The following code could be adapted to write data from devices using the set command as 
in client.set("string key", "string val", function(err,reply){});